+++ include/vrv/functorparams.h
@@ -41,6 +41,7 @@ class Measure;
 class MeasureAligner;
 class Mensur;
 class MeterSig;
+class MeterSigGrp;
 class MRpt;
 class Object;
 class Page;
@@ -575,6 +576,7 @@ public:
     {
         m_tupletNum = tupletNum;
         m_drawingNumPos = STAFFREL_basic_NONE;
+        m_horizontalMargin = 0;
         m_verticalMargin = 0;
         m_ignoreCrossStaff = false;
         m_yRel = 0;
@@ -582,6 +584,7 @@ public:
 
     TupletNum *m_tupletNum;
     data_STAFFREL_basic m_drawingNumPos;
+    int m_horizontalMargin;
     int m_verticalMargin;
     bool m_ignoreCrossStaff;
     int m_yRel;
@@ -649,11 +652,14 @@ public:
  * member 5: the list of staffN in the top-level scoreDef
  * member 6: the bounding box in the previous aligner
  * member 7: the upcoming bounding boxes (to be used in the next aligner)
- * member 8: the Doc
- * member 9: the Functor for redirection to the MeasureAligner
- * member 10: the end Functor for redirection
- * member 11: current aligner that is being processed
- * member 12: preceeding aligner that was handled before
+ * member 8: list of types to include
+ * member 9: list of types to exclude
+ * member 10: list of tie endpoints for the current measure
+ * member 11: the Doc
+ * member 12: the Functor for redirection to the MeasureAligner
+ * member 13: the end Functor for redirection
+ * member 14: current aligner that is being processed
+ * member 15: preceeding aligner that was handled before
  **/
 
 class AdjustXPosParams : public FunctorParams {
@@ -671,6 +677,7 @@ public:
         m_functorEnd = functorEnd;
         m_currentAlignment.Reset();
         m_previousAlignment.Reset();
+        m_measureTieEndpoints.clear();
     }
     int m_minPos;
     int m_upcomingMinPos;
@@ -682,6 +689,7 @@ public:
     std::vector<BoundingBox *> m_upcomingBoundingBoxes;
     std::vector<ClassId> m_includes;
     std::vector<ClassId> m_excludes;
+    std::vector<std::pair<LayerElement *, LayerElement *>> m_measureTieEndpoints;
     Doc *m_doc;
     Functor *m_functor;
     Functor *m_functorEnd;
@@ -954,14 +962,14 @@ public:
         m_currentRealTimeSeconds = 0.0;
         m_maxCurrentScoreTime = 0.0;
         m_maxCurrentRealTimeSeconds = 0.0;
-        m_currentTempo = 120;
+        m_currentTempo = 120.0;
         m_tempoAdjustment = 1.0;
     }
     double m_currentScoreTime;
     double m_currentRealTimeSeconds;
     double m_maxCurrentScoreTime;
     double m_maxCurrentRealTimeSeconds;
-    int m_currentTempo;
+    double m_currentTempo;
     double m_tempoAdjustment;
 };
 
@@ -987,14 +995,14 @@ public:
         m_currentMensur = NULL;
         m_currentMeterSig = NULL;
         m_notationType = NOTATIONTYPE_cmn;
-        m_currentTempo = 120;
+        m_currentTempo = 120.0;
     }
     double m_currentScoreTime;
     double m_currentRealTimeSeconds;
     Mensur *m_currentMensur;
     MeterSig *m_currentMeterSig;
     data_NOTATIONTYPE m_notationType;
-    int m_currentTempo;
+    double m_currentTempo;
 };
 
 //----------------------------------------------------------------------------
@@ -1434,26 +1442,21 @@ public:
  * member 2: the maximum position
  * member 3: the timespanning interface
  * member 4: the class Ids to keep
- * member 5: the slur for finding ties (too specific, to be refactored)
- * member 6: the ties we need to consider (too specific, to be refactored)
  **/
 
 class FindSpannedLayerElementsParams : public FunctorParams {
 public:
-    FindSpannedLayerElementsParams(TimeSpanningInterface *interface, Slur *slur)
+    FindSpannedLayerElementsParams(TimeSpanningInterface *interface)
     {
         m_interface = interface;
         m_minPos = 0;
         m_maxPos = 0;
-        m_slur = slur;
     }
     std::vector<LayerElement *> m_elements;
     int m_minPos;
     int m_maxPos;
     TimeSpanningInterface *m_interface;
     std::vector<ClassId> m_classIds;
-    Slur *m_slur;
-    std::vector<FloatingPositioner *> m_ties;
 };
 
 //----------------------------------------------------------------------------
@@ -1475,7 +1478,7 @@ using MIDINoteSequence = std::list<MIDINote>;
  * member 1: int: the midi track number
  * member 3: double: the score time from the start of the music to the start of the current measure
  * member 4: int: the semi tone transposition for the current track
- * member 5: int with the current tempo
+ * member 5: double with the current tempo
  * member 6: expanded notes due to ornaments and tremolandi
  **/
 
@@ -1488,7 +1491,7 @@ public:
         m_midiTrack = 1;
         m_totalTime = 0.0;
         m_transSemi = 0;
-        m_currentTempo = 120;
+        m_currentTempo = 120.0;
         m_functor = functor;
     }
     smf::MidiFile *m_midiFile;
@@ -1496,7 +1499,7 @@ public:
     int m_midiTrack;
     double m_totalTime;
     int m_transSemi;
-    int m_currentTempo;
+    double m_currentTempo;
     std::map<Note *, MIDINoteSequence> m_expandedNotes;
     Functor *m_functor;
 };
@@ -1521,16 +1524,16 @@ public:
     {
         m_scoreTimeOffset = 0.0;
         m_realTimeOffsetMilliseconds = 0;
-        m_currentTempo = 120;
+        m_currentTempo = 120.0;
         m_functor = functor;
     }
     std::map<double, double> realTimeToScoreTime;
     std::map<double, std::vector<std::string>> realTimeToOnElements;
     std::map<double, std::vector<std::string>> realTimeToOffElements;
-    std::map<double, int> realTimeToTempo;
+    std::map<double, double> realTimeToTempo;
     double m_scoreTimeOffset;
     double m_realTimeOffsetMilliseconds;
-    int m_currentTempo;
+    double m_currentTempo;
     Functor *m_functor;
 };
 
@@ -1967,18 +1970,20 @@ public:
 
 class ReplaceDrawingValuesInStaffDefParams : public FunctorParams {
 public:
-    ReplaceDrawingValuesInStaffDefParams(
-        Clef const *clef, KeySig const *keySig, Mensur const *mensur, MeterSig const *meterSig)
+    ReplaceDrawingValuesInStaffDefParams(Clef const *clef, KeySig const *keySig, Mensur const *mensur,
+        MeterSig const *meterSig, MeterSigGrp const *meterSigGrp)
     {
         m_clef = clef;
         m_keySig = keySig;
         m_mensur = mensur;
         m_meterSig = meterSig;
+        m_meterSigGrp = meterSigGrp;
     }
     Clef const *m_clef;
     KeySig const *m_keySig;
     Mensur const *m_mensur;
     MeterSig const *m_meterSig;
+    MeterSigGrp const *m_meterSigGrp;
 };
 
 //----------------------------------------------------------------------------
@@ -2195,21 +2200,22 @@ public:
  * member 4: bool the flag for indicating if apply to all or not
  **/
 
+enum StaffDefRedrawFlags {
+    REDRAW_CLEF = 0x1,
+    REDRAW_KEYSIG = 0x2,
+    REDRAW_MENSUR = 0x4,
+    REDRAW_METERSIG = 0x8,
+    REDRAW_METERSIGGRP = 0x10,
+    // all flags
+    REDRAW_ALL = REDRAW_CLEF | REDRAW_KEYSIG | REDRAW_MENSUR | REDRAW_METERSIG | REDRAW_METERSIGGRP,
+    //
+    FORCE_REDRAW = 0x100
+};
+
 class SetStaffDefRedrawFlagsParams : public FunctorParams {
 public:
-    SetStaffDefRedrawFlagsParams()
-    {
-        m_clef = false;
-        m_keySig = false;
-        m_mensur = false;
-        m_meterSig = false;
-        m_applyToAll = false;
-    }
-    bool m_clef;
-    bool m_keySig;
-    bool m_mensur;
-    bool m_meterSig;
-    bool m_applyToAll;
+    SetStaffDefRedrawFlagsParams() { m_redrawFlags = 0; }
+    int m_redrawFlags;
 };
 
 //----------------------------------------------------------------------------