+++ include/vrv/vrvdef.h
@@ -36,7 +36,7 @@ namespace vrv {
 //----------------------------------------------------------------------------
 
 #define VERSION_MAJOR 3
-#define VERSION_MINOR 5
+#define VERSION_MINOR 6
 #define VERSION_REVISION 0
 // Adds "-dev" in the version number - should be set to false for releases
 #define VERSION_DEV true
@@ -103,6 +103,7 @@ enum ClassId {
     MEASURE_ALIGNER,
     MENSUR_ATTR,
     METERSIG_ATTR,
+    METERSIGGRP,
     PAGE,
     PAGES,
     SCORE,
@@ -169,6 +170,7 @@ enum ClassId {
     GLISS,
     HAIRPIN,
     HARM,
+    LV,
     MORDENT,
     MNUM,
     OCTAVE,
@@ -337,7 +339,7 @@ typedef std::vector<LedgerLine> ArrayOfLedgerLines;
 
 typedef std::vector<TextElement *> ArrayOfTextElements;
 
-typedef std::map<Staff *, std::set<int>> MapOfNoteLocs;
+typedef std::map<Staff *, std::multiset<int>> MapOfNoteLocs;
 
 typedef std::map<Staff *, std::set<int>> MapOfDotLocs;