+++ CHANGELOG.md
@@ -1,14 +1,21 @@
 # Changelog
 
 ## [unreleased]
+
+## [3.5.0] - 2021-07-22
+* Support for `lv` (@eNote-GmbH)
 * Support for additive meters (@eNote-GmbH)
 * Support for `bTrem` (bowed tremolos) in the MIDI output (@eNote-GmbH)
+* Support for `@enclose` on articulation and fermata (@eNote-GmbH)
 * Improved rendering of cross-staff slurs (@eNote-GmbH)
 * Improved octave line positions with `@endid` (@eNote-GmbH)
 * Improved layout algorithm with nested layer content arrangement
 * Improved rendering to prevent rendering barlines for the invisible measure (@eNote-GmbH)
 * Improved rendering of octave brackets ending (@eNote-GmbH)
+* Function getTimeForElement includes `measure` and `chord` elements (@gitelicious)
 * Option --breaks-no-widow to prevent a single measure on last page (@eNote-GmbH)
+* Option --ligature-as-bracket to render ligatures as bracket instead of original notation
+* Option --lyric-verse-collapse to collapse empty lyric lines
 * Build option NO_MXL_SUPPORT to disable compressed MusicXML file support
 * Making Verovio thread safe (@eNote-GmbH)
 * Switch from XSLT to Python for font generation (@eNote-GmbH)